My idea for how to fix sign in and processing issues.
